fix(qov-860): Set integer as u64 by default (#37)
* fix(qov-860): Force uint64 as u64

Despite setting the good format in the openapi, the generated rust type
leads to i32, such as:
```
root_volume_size:
  type: integer
  description: System volume disk size. (in bytes)
  format: uint64
  nullable: true
```

To resolve the issue, force map the u64 type in the openapi generation
command with `--type-mappings=integer+uint64=u64`

* fix(qov-860): Force uint32 as u64

+++ src/apis/node_types_api.rs
@@ -26,7 +26,7 @@ pub async fn list_node_types(
     region: &str,
     include_disabled_types: bool,
     page: Option<i32>,
-    page_size: Option<i32>,
+    page_size: Option<u64>,
 ) -> Result<models::ScalewayPeriodRdbPeriodV1PeriodListNodeTypesResponse, Error<ListNodeTypesError>>
 {
     // add a prefix to parameters to efficiently prevent name collisions
