+++ src/models/scaleway_period_baremetal_period_v1_period_memory.rs
@@ -15,13 +15,13 @@ use serde::{Deserialize, Serialize};
 pub struct ScalewayPeriodBaremetalPeriodV1PeriodMemory {
     /// Capacity of the memory in bytes. (in bytes)
     #[serde(rename = "capacity", skip_serializing_if = "Option::is_none")]
-    pub capacity: Option<i32>,
+    pub capacity: Option<u64>,
     /// Type of the memory.
     #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
     pub r#type: Option<String>,
     /// Frequency of the memory in MHz.
     #[serde(rename = "frequency", skip_serializing_if = "Option::is_none")]
-    pub frequency: Option<i32>,
+    pub frequency: Option<u64>,
     /// True if the memory is an error-correcting code memory.
     #[serde(rename = "is_ecc", skip_serializing_if = "Option::is_none")]
     pub is_ecc: Option<bool>,
